メイン処理のうち単純表示にしか使わない部分を単純表示処理に移行
